// src/spec/threatgile.generated.ts
export interface Threagile {
// ...
  trust_boundaries?: {
    [k: string]: {
      /**
       * ID
       */
      id: string;
      /**
       * Description
       */
      description: string | null;
      /**
       * Type
       */
      type:
        | "network-on-prem"
        | "network-dedicated-hoster"
        | "network-virtual-lan"
        | "network-cloud-provider"
        | "network-cloud-security-group"
        | "network-policy-namespace-isolation"
        | "execution-environment";
      /**
       * Tags
       */
      tags?: string[] | null;
      /**
       * Technical assets inside
       */
      technical_assets_inside: string[] | null;
      /**
       * Trust boundaries nested
       */
      trust_boundaries_nested: string[] | null;
      [k: string]: unknown;
    };
  };
// ...
}